rag_clinical_notes/requirements.txt
# PDF processing
pdfplumber==0.11.0
reportlab==4.2.2

# LangChain stack
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
langchain-huggingface==0.0.3
langchain-experimental==0.0.64

# Embeddings + re-ranking
sentence-transformers==3.0.1

# Vector store
chromadb==0.5.5

# Guardrails
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.5
# Run once after install: python -m spacy download en_core_web_sm

# Evaluation
ragas==0.1.21
datasets==2.20.0

# Serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# MLflow (shared with root project)
mlflow==2.13.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
